+++ Projects/Feature/FindCompanyFeature/Sources/FindCompanyComponent.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+import HomeFeatureInterface
+import CompaniesDomainInterface
+import NeedleFoundation
+
+public protocol FindCompanyDependency: Dependency {
+    var fetchStudentCompanyListUseCase: any FetchStudentCompanyListUseCase { get }
+}
+
+public final class FindCompanyComponent: Component<FindWorkSpaceDependency>, FindCompanyFactory {
+    public func makeView() -> some View {
+        FindCompanyView(
+            viewModel: .init(
+                fetchStudentCompanyListUseCase: dependency.fetchStudentCompanyListUseCase
+            )
+        )
+    }
+}
