--- Projects/Feature/HomeFeature/Sources/FindWorkSpace/FindWorkSpaceComponent.swift
@@ -1,16 +1,0 @@
-import SwiftUI
-import HomeFeatureInterface
-import CompaniesDomainInterface
-import NeedleFoundation
-
-public protocol FindWorkSpaceDependency: Dependency {
-    var fetchStudentCompanyListUseCase: any FetchStudentCompanyListUseCase { get }
-}
-
-public final class FindWorkSpaceComponent: Component<FindWorkSpaceDependency>, FindWorkSpaceFactory {
-    public func makeView() -> some View {
-        FindWorkSpaceView(
-            viewModel: .init(fetchStudentCompanyListUseCase: dependency.fetchStudentCompanyListUseCase)
-        )
-    }
-}
